#hotfix 포트폴리오 썸네일 이미지 출력안되는 버그 수정

+++ src/main/java/com/casker/portfolio/controller/FileController.java
@@ -42,7 +42,7 @@ public class FileController {
 	 */
 	@ResponseBody
 	@RequestMapping("/portfolio/{portfolioNo}/{imageType}")
-	public void editPassword(HttpServletResponse response, @PathVariable int portfolioNo, @PathVariable String imageType) throws Exception {
+	public void editPassword(HttpServletResponse response, @PathVariable long portfolioNo, @PathVariable String imageType) throws Exception {
 		
 		Portfolio portfolio = portfolioService.getPortfolioDetail(portfolioNo);
